Worked-around crash that would happen when many files are selected
What happens is the shell extension will try (de)compressing as
Kosinski. This is because Windows is braindead and can't tell the
difference between 'Kosinski/Compress' and 'Enigma/Compress'... but only
when *a lot* of files are selected (50 is enough to trigger it).

So to work around it, no two options have the same name.